pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.34",
    "python-dotenv>=1.0.0",
    "fastapi>=0.128.0",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=2.0.0",
    "langgraph>=1.0.7",
    "mcp>=1.25.0",
    "langchain-mcp-adapters>=0.1.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
    "psycopg2-binary>=2.9.9",
    "PyJWT[crypto]>=2.8.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
    "reportlab>=4.4.10",
]
